tutorials/ner_drugs/scripts/deployment/requirements.txt
python-dotenv==0.19.2
wandb==0.12.10
spacy==3.2.1
typer==0.4.0